Load all files into memory on load
Up to now, all the information is loaded from files
on every query. It isn't the most elegant way of working
with data, since it's slow. But also it brings several shortcomings
on multiple planes. One example is the boilerplate that we need to
go through every time a new command is created.